requirements.txt
aiohttp
websockets
pyyaml
asyncio-mqtt
